make flags unsigned long + other changes

+++ src/num/mdfft.h
@@ -7,12 +7,12 @@
 
 
 extern void md_fft2(unsigned int N, const long dims[N],
-	unsigned int flags, unsigned int dirs,  
+	unsigned long flags, unsigned long dirs,
 	const long ostr[N], complex float* dst,
 	const long istr[N], const complex float* in);
 
 extern void md_fft(unsigned int N, const long dims[N],
-		unsigned int flags, unsigned int dirs,
+		unsigned long flags, unsigned long dirs,
 		complex float* dst, const complex float* in);
